Default stylin'
Flex boxes turned out to be the easiest way to get a good layout. Since each box is 300px by 300px and we use "justify-content: center", this will work great from small devices up to at least laptop screens. On my laptop, I can have the chrome dev tools open to fiddle with it, and the boxes will be neatly in a 2 by 2 grid.